requirements.txt
python-telegram-bot
motor
pymongo
dnspython